### 2025-04-17 #700 @feddelegrand7
rlowdb package submission
Submitting Author Name: Mohamed El Fodil Ihaddaden
Submitting Author Github Handle: @feddelegrand7
Repository:  https://github.com/feddelegrand7/rlowdb
Version submitted: 0.2.0
Submission type: Standard

Version accepted: TBD

---

```
Package: rlowdb
Title: Lightweight JSON-Based Database
Version: 0.2.0
    person(given = "Mohamed El Fodil",
           family = "Ihaddaden",
           email = "ihaddaden.fodeil@gmail.com"),
     person(given = "lowdb developers",
            role = c("ctb", "cph"),
            comment = "developers of the lowdb package"))
Description: The goal of 'rlowdb' is to provide a lightweight, file-based JSON database.
    Inspired by 'LowDB' in 'JavaScript', it generates an intuitive interface for
    storing, retrieving, updating, and querying structured data without requiring a full-fledged database system.
    Ideal for prototyping, small-scale applications, and lightweight data management needs.
    cli (>= 3.6.4),
    purrr (>= 1.0.2),
    R6 (>= 2.5.1),
    rlang (>= 1.1.3),
    yyjsonr (>= 0.1.20)
```

The package is a lightweight, JSON-based database for R. It provides a simple and efficient way to store, retrieve, update, and delete structured data without the need for a full database system.

Any R developer that needs to work with a persistent data source.
At the moment, to the best of my knowledge, there's no equivalent in the R ecosystem.
Yes
