+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-04-16T11:19Z
+last_built: 2025-04-17T10:07Z
 urls:
   reference: https://feddelegrand7.github.io/rlowdb/reference
   article: https://feddelegrand7.github.io/rlowdb/articles
