### 2025-04-18 #700 @maurolepore
@feddelegrand7 thanks a lot for sharing your work with rOpenSci.

The package looks really neat. Let me discuss the scope with the editorial board. I'll come back ASAP.
