### 2025-04-20 #700 @maurolepore
@feddelegrand7 from the information we have we think the package might be in scope for the category "database software bindings". We don't see the fit for the other categories.

> **database software bindings**: Bindings and wrappers for generic database APIs (Example: [rrlite](htttps://github.com/ropensci/software-review/issues/6))
-- https://devdevguide.netlify.app/softwarereview_policies.html#package-categories

What do you think? Is there a reason why you did not select that category?

If you can make a case for the package to fit in that category, then please edit the top comment to reflect that. You'll need to explain why it fits in that category.

Also, note that before we consider the package for review [all of these checks should pass](https://github.com/ropensci/software-review/issues/700#issuecomment-2812380330).
