### 2025-04-22 #700 @feddelegrand7
Hello @maurolepore thanks for your feedback. You're absolutely right. Database software bindings is a good fit, I just didn't see it ^^

I'll modify the package so that it passes the checks and come back to you.
