### 2025-04-22 #700 @feddelegrand7
@maurolepore any resources on how to create the above mentioned files? codemeta.json, contributing?
