### 2025-04-25 487828a
adding the auto_reload logic
