+++ R/main.R
@@ -25,6 +25,10 @@ rlowdb <- R6::R6Class(
     #' @param auto_commit whether to update the DB automatically each time
     #' there's an insertion, an update or a deletion. Defaults to TRUE.s
     #' Note that you can use the `commit` method to update the DB manually.
+    #' @param auto_reload if set to TRUE, will always re-read the underlying JSON time before
+    #' running the methods. Useful, if you have many people working simultaneously on the same JSON file.
+    #' FALSE by default for performance purpose. Note that you can manually reload your data to the latest
+    #' stand using the `reload` method.
     #' @param verbose If TRUE, will print informative messages to the console.
     #' Defaults to FALSE
     #' @param pretty Use pretty = FALSE for compact JSON,
@@ -34,6 +38,7 @@ rlowdb <- R6::R6Class(
     file_path,
     default_values = list(),
     auto_commit = TRUE,
+    auto_reload = FALSE,
     verbose = FALSE,
     pretty = FALSE
     ) {
@@ -55,6 +60,7 @@ rlowdb <- R6::R6Class(
 
       private$.file_path <- file_path
       private$.auto_commit <- auto_commit
+      private$.auto_reload <- auto_reload
       private$.verbose <- verbose
       private$.default_values <- default_values
       private$.pretty <- pretty
@@ -66,6 +72,18 @@ rlowdb <- R6::R6Class(
       private$.write_data()
     },
 
+    #' @description Refreshes the in-memory database by reading the current state from the JSON file
+    #' @return rlowdb object invisibly for chaining
+    reload = function() {
+      private$.read_data()
+
+      if (private$.verbose) {
+        cli::cli_alert_success("Database reloaded from {private$.file_path}")
+      }
+
+      invisible(self)
+    },
+
     #' @description Retrieve all stored data.
     #' @return A list containing all database records.
     #' @examples
@@ -75,6 +93,11 @@ rlowdb <- R6::R6Class(
     #' unlink("database.json")
     #
     get_data = function() {
+
+      if (private$.auto_reload) {
+        self$reload()
+      }
+
       private$.data
     },
 
@@ -88,6 +111,11 @@ rlowdb <- R6::R6Class(
     #' unlink("database.json")
 
     get_data_collection = function(collection) {
+
+      if (private$.auto_reload) {
+        self$reload()
+      }
+
       if (!collection %in% names(private$.data)) {
         rlang::abort(sprintf("Error: Collection '%s' does not exist.", collection))
       }
@@ -107,6 +135,10 @@ rlowdb <- R6::R6Class(
     #' unlink("database.json")
 
     get_data_key = function(collection, key) {
+
+      if (private$.auto_reload) {
+        self$reload()
+      }
 
       if (!collection %in% names(private$.data)) {
         rlang::abort(sprintf("Error: Collection '%s' does not exist.", collection))
@@ -131,8 +163,13 @@ rlowdb <- R6::R6Class(
     #' )
     #' db$insert("users", list(id = 1, name = "Alice"))
     #' unlink("database.json")
+    #' @return rlowdb object invisibly for chaining
     #'
     insert = function(collection, record) {
+
+      if (private$.auto_reload) {
+        self$reload()
+      }
 
       if (!is.list(record) || is.null(names(record)) || !all(nzchar(names(record)))) {
         rlang::abort("Error: 'record' must be a named list with valid field names.")
@@ -176,6 +213,11 @@ rlowdb <- R6::R6Class(
     #' unlink("database.json")
     #'
     find = function(collection, key, value) {
+
+      if (private$.auto_reload) {
[…]
