+++ man/rlowdb.Rd
@@ -50,7 +50,6 @@ db <- rlowdb$new("database.json", default_values = list(
 db$insert("users", list(id = 1, name = "Alice"))
 unlink("database.json")
 
-
 ## ------------------------------------------------
 ## Method `rlowdb$find`
 ## ------------------------------------------------
@@ -69,7 +68,6 @@ db <- rlowdb$new("database.json")
 db$insert("users", list(id = 1, name = "Alice"))
 db$update("users", "id", 1, list(name = "Alice Updated"))
 unlink("database.json")
-
 
 ## ------------------------------------------------
 ## Method `rlowdb$upsert`
@@ -90,7 +88,6 @@ db$insert("users", list(id = 1, name = "Alice"))
 db$delete("users", "id", 1)
 db$get_data()
 unlink("database.json")
-
 
 ## ------------------------------------------------
 ## Method `rlowdb$query`
@@ -154,6 +151,7 @@ db$insert("consumers", list(name = "Teta", age = 22))
 db$clear("users")
 db$get_data()
 unlink("database.json")
+
 
 ## ------------------------------------------------
 ## Method `rlowdb$count`
@@ -221,6 +219,7 @@ db$transaction(function() {
 db$count("users")
 unlink("database.json")
 
+
 ## ------------------------------------------------
 ## Method `rlowdb$search`
 ## ------------------------------------------------
@@ -250,6 +249,7 @@ db$bulk_insert("users", list(
 db$count("users")
 unlink("database.json")
 
+
 ## ------------------------------------------------
 ## Method `rlowdb$rename_collection`
 ## ------------------------------------------------
@@ -264,7 +264,6 @@ db$list_collections()
 db$rename_collection("users", "customers")
 db$list_collections()
 unlink("database.json")
-
 
 ## ------------------------------------------------
 ## Method `rlowdb$list_keys`
@@ -312,6 +311,7 @@ db$insert_default_values("users", list(role = "guest", active = TRUE))
 db$insert_default_values("users", list(role = "guest", active = TRUE), replace_existing = TRUE)
 
 unlink("database.json")
+
 
 ## ------------------------------------------------
 ## Method `rlowdb$clone_collection`
@@ -364,6 +364,7 @@ try(db$insert("users", list(id = "1", name = "")))
 \itemize{
 \item \href{#method-rlowdb-new}{\code{rlowdb$new()}}
 \item \href{#method-rlowdb-commit}{\code{rlowdb$commit()}}
+\item \href{#method-rlowdb-reload}{\code{rlowdb$reload()}}
 \item \href{#method-rlowdb-get_data}{\code{rlowdb$get_data()}}
 \item \href{#method-rlowdb-get_data_collection}{\code{rlowdb$get_data_collection()}}
 \item \href{#method-rlowdb-get_data_key}{\code{rlowdb$get_data_key()}}
@@ -412,6 +413,7 @@ If the file does not exist, an empty database is created.
   file_path,
   default_values = list(),
   auto_commit = TRUE,
+  auto_reload = FALSE,
   verbose = FALSE,
   pretty = FALSE
 )}\if{html}{\out{</div>}}
@@ -432,6 +434,11 @@ Default is an empty list (`list()`).}
 there's an insertion, an update or a deletion. Defaults to TRUE.s
 Note that you can use the `commit` method to update the DB manually.}
 
+\item{\code{auto_reload}}{if set to TRUE, will always re-read the underlying JSON time before
[…]
