### 2025-05-04 #700 @maurolepore
This marks the end of my EiC rotation and leaves a note for the next one.

* The pakcage seems in scope but has not been announced to editors because pkgcheck detects some issues that need to be address before we proceed.
* More details here: https://github.com/ropensci/software-review/issues/700#issuecomment-2817159421

@feddelegrand7 thanks a lot for sharing your work with rOpenSci.
