+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-04-25T12:05Z
+last_built: 2025-05-15T11:34Z
 urls:
   reference: https://feddelegrand7.github.io/rlowdb/reference
   article: https://feddelegrand7.github.io/rlowdb/articles
