### 2026-02-07 #700 @emilyriederer
@feddelegrand7 I have recently started my EiC rotation, taking over for Jeff, and am conducting a review of stale issues.

To reiterate, this package may be in scope for rOpenSci, but prior to moving forward the failing checks would need to be addressed. As it has been >1 month since we've been in touch, we will close this issue for now. If you decide you'd like to proceed with a review, please feel free to resubmit. Thank you for your interest in rOpenSci!
